+++ Source/ProjectRimFactory/Common/HarmonyPatches/Patch_GenRecipe_MakeRecipeProducts_foodPoisoning.cs
@@ -1,5 +1,4 @@
 using HarmonyLib;
-using ProjectRimFactory.Drones;
 using ProjectRimFactory.SAL3.Things.Assemblers;
 using RimWorld;
 using System;
@@ -28,7 +27,7 @@ namespace ProjectRimFactory.Common.HarmonyPatches
      * use Transpiler to modify the code.
      */
     [HarmonyPatch]
-    public static class Patch_GenRecipe_foodPoisoning
+    public static class Patch_GenRecipe_MakeRecipeProducts_foodPoisoning
     {
         /// <summary>
         /// Doing this should find the inner iterator class no matter how the compiler calls it.
@@ -75,8 +74,8 @@ namespace ProjectRimFactory.Common.HarmonyPatches
                     yield return new CodeInstruction(OpCodes.Ldfld, AccessTools.Field(hiddenClass, "billGiver"));
                     // Call our UsingSpaceCooker method
                     yield return new CodeInstruction(OpCodes.Call, AccessTools.Method(
-                        typeof(Patch_GenRecipe_foodPoisoning),
-                        nameof(Patch_GenRecipe_foodPoisoning.UsingSpaceCooker)));
+                        typeof(Patch_GenRecipe_MakeRecipeProducts_foodPoisoning),
+                        nameof(Patch_GenRecipe_MakeRecipeProducts_foodPoisoning.UsingSpaceCooker)));
                     // If that returned true, skip past the original condition
                     yield return new CodeInstruction(OpCodes.Brtrue_S, (Label)instruction.operand);
                     codeHasStoredCompPoisonable = false;
@@ -97,8 +96,8 @@ namespace ProjectRimFactory.Common.HarmonyPatches
                     yield return new CodeInstruction(OpCodes.Ldarg_0);
                     yield return new CodeInstruction(OpCodes.Ldfld, AccessTools.Field(hiddenClass, "billGiver"));
                     yield return new CodeInstruction(OpCodes.Call, AccessTools.Method(
-                        typeof(Patch_GenRecipe_foodPoisoning),
-                        nameof(Patch_GenRecipe_foodPoisoning.GetRoomOfPawnOrGiver)));
+                        typeof(Patch_GenRecipe_MakeRecipeProducts_foodPoisoning),
+                        nameof(Patch_GenRecipe_MakeRecipeProducts_foodPoisoning.GetRoomOfPawnOrGiver)));
 
                     continue; // Don't emit the original instruction.
                 }
@@ -137,80 +136,6 @@ namespace ProjectRimFactory.Common.HarmonyPatches
             }
             return pawn.GetRoom(allowedRegionTypes);
         }
-    }
-
-    [HarmonyPatch(typeof(QualityUtility), "GenerateQualityCreatedByPawn", new Type[] { typeof(Pawn), typeof(SkillDef) })]
-    class Patch_GenRecipe_GenerateQualityCreatedByPawn
-    {
-        static bool Prefix(ref QualityCategory __result, Pawn pawn, SkillDef relevantSkill)
-        {
-            ISetQualityDirectly isqd = PatchStorageUtil.Get<ISetQualityDirectly>(pawn.Map, pawn.Position);
-            if (isqd != null)
-            {
-                __result = isqd.GetQuality(relevantSkill);
-                return false;
-            }
-            return true;
-        }
-
-        //Prevent Biotech(Mech Changes) from interfering with Drone Skills
-        static IEnumerable<CodeInstruction> Transpiler(IEnumerable<CodeInstruction> instructions)
-        {
-
-            bool foundReplaceMarker = false;
-            bool repacedCheck = false;
-            int cnt = 0;
-            foreach (var instruction in instructions)
-            {
-                cnt++;
-                //Search for IL_0000: ldarg.0 -> Loading the Pawn
-                if (instruction.opcode == OpCodes.Ldarg_0 && !repacedCheck)
-                {
-                    repacedCheck = true;
-                    foundReplaceMarker = true;
-                    cnt = 0;
-                }
-                //remove IL_0001: callvirt instance class Verse.RaceProperties Verse.Pawn::get_RaceProps()
-                if (instruction.opcode == OpCodes.Callvirt && foundReplaceMarker && cnt == 1)
-                {
-                    continue;
-                }
-                if (instruction.opcode == OpCodes.Callvirt && foundReplaceMarker && cnt == 2)
-                {
-                    //Replace IL_0006: callvirt instance bool Verse.RaceProperties::get_IsMechanoid()
-                    //with a call to Patch_GenRecipe_GenerateQualityCreatedByPawn:IsMechanoid(Pawn pawn)
-                    instruction.operand = AccessTools.Method(
-                        typeof(Patch_GenRecipe_GenerateQualityCreatedByPawn),
-                        nameof(Patch_GenRecipe_GenerateQualityCreatedByPawn.IsMechanoid), new[] { typeof(Pawn)});
-                    foundReplaceMarker = false;
-                }
-
-
-                yield return instruction;
-            }
-
-        }
-
-        /// <summary>
-        /// if a Pawn is a Mechanoid then it's skills will be ignored by QualityUtility:GenerateQualityCreatedByPawn
-        /// This prevents Drones from being detected as Mechanoid
-        /// </summary>
-        /// <param name="pawn"></param>
-        /// <returns></returns>
-        public static bool IsMechanoid(Pawn pawn)
-        {
-            if (pawn is Pawn_Drone)
-            {
-                return false;
-            }
-            return pawn.RaceProps.IsMechanoid;
-        }
-
-    }
-
-    interface ISetQualityDirectly
-    {
-        QualityCategory GetQuality(SkillDef relevantSkill);
     }
 
 }
